agent-under-observation/requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=1.0.0
langgraph-checkpoint-sqlite>=2.0.0
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
typing-extensions>=4.0.0
psycopg2-binary>=2.9.0
duckduckgo-search>=6.0.0
faker>=24.0.0
